Include compile commands for targets marked manual
tools/gen_compilation_database.py is used to generate a JSON file that
tells other tools how to compile each .cc file. It does this by calling
`bazel build` with a list of target patterns. When bazel expands
`//source/...` and such, it excludes targets marked "manual". Since
clang-tidy and such depend on the database to be complete, do an extra
query to find those manual targets (by default) and include them in the
database.

Signed-off-by: Alex Konradi <akonradi@google.com>